chore: extract inline tests to sibling tests.rs files (db crate) (#413)
Mirrors the server-side cleanup in PR #409: moves the inline
`#[cfg(test)] mod tests` blocks out of five db production modules into
sibling `tests.rs` files following the established
`db/src/books/tests.rs` pattern.

- db/src/browse.rs       (623 -> 201 lines) — extracted ~423 lines of tests
- db/src/settings.rs     (630 -> 252 lines) — extracted two inline blocks
                                              (test_helpers + tests); the
                                              module-local LibraryEnvGuard
                                              moved into settings/tests.rs
                                              since it had no other consumer
- db/src/progress.rs     (408 -> 172 lines) — extracted ~237 lines of tests
- db/src/indexer.rs      (822 -> 408 lines) — extracted ~413 lines of tests
- db/src/thumbs.rs       (396 -> 268 lines) — extracted ~129 lines of tests

Existing cross-crate helpers (indexed, CoversTempDir, seed_discovery_fixture,
etc.) already lived in db/src/test_support.rs — no new hoisting was needed.

Also anchors the three runtime-cache .gitignore entries (`covers/`, `thumbs/`,
`data/`) to repo-root with a leading `/` so the new sibling `db/src/thumbs/`
test directory isn't ignored. The runtime caches still ignored at the root.

Production-module public API unchanged; all 454 db tests pass under the
same names.

Closes #413
