+++ .gitignore
@@ -7,11 +7,11 @@ target
 *.db-wal
 
 # Dev-server cover cache (F0.1 writes extracted covers to ./covers by default)
-covers/
+/covers/
 # Dev-server thumbnail WebP cache (F1.2)
-thumbs/
+/thumbs/
 # F2.3 HLS transcode cache (OMNIBUS_DATA_DIR default = ./data; segments + concat.txt + .progress)
-data/
+/data/
 
 # These are backup files generated by rustfmt
 **/*.rs.bk
