+++ db/src/progress/tests.rs
@@ -1,0 +1,240 @@
+//! Unit tests for the `progress` module — `upsert_progress`
+//! roundtrip + last-write-wins + per-user/format isolation,
+//! `BookNotFound` variant, `get_progress` empty-state, and
+//! `record_session` per-format dispatch with unknown-uuid skip.
+
+use super::*;
+use crate::{init_db, replace_books};
+use omnibus_shared::EbookMetadata;
+
+async fn seed(pool: &SqlitePool, library: &str, title: &str) -> (i64, String) {
+    replace_books(
+        pool,
+        library,
+        vec![crate::ebook::IndexedBook {
+            metadata: EbookMetadata {
+                filename: format!("{title}.epub").to_lowercase(),
+                title: Some(title.to_string()),
+                ..Default::default()
+            },
+            cover: None,
+            mtime_epoch: 0,
+            size_bytes: 0,
+        }],
+    )
+    .await
+    .expect("seed book");
+    let books = crate::list_books(pool, library).await.unwrap();
+    let book = books
+        .into_iter()
+        .find(|b| b.title.as_deref() == Some(title))
+        .unwrap();
+    (book.id, book.unique_identifier.clone().unwrap())
+}
+
+async fn seed_user(pool: &SqlitePool, name: &str) -> i64 {
+    sqlx::query_scalar::<_, i64>(
+        "INSERT INTO users (username, password_hash, is_admin, can_upload, can_edit, can_download)
+         VALUES (?, '!x', 0, 0, 0, 1) RETURNING id",
+    )
+    .bind(name)
+    .fetch_one(pool)
+    .await
+    .unwrap()
+}
+
+#[tokio::test]
+async fn upsert_round_trips_epub_position() {
+    let pool = init_db("sqlite::memory:").await.unwrap();
+    let user = seed_user(&pool, "alice").await;
+    let (_, uuid) = seed(&pool, "/lib", "Book A").await;
+    let upd = ProgressUpdate {
+        book_uuid: uuid.clone(),
+        format: ProgressFormat::Epub,
+        epub_cfi: Some("epubcfi(/6/4!/4/2/1:0)".into()),
+        audio_position_seconds: None,
+    };
+    let saved = upsert_progress(&pool, user, &upd).await.unwrap();
+    assert_eq!(saved.book_uuid, uuid);
+    assert_eq!(saved.format, ProgressFormat::Epub);
+    assert_eq!(saved.epub_cfi.as_deref(), Some("epubcfi(/6/4!/4/2/1:0)"));
+    assert!(saved.updated_at > 0);
+
+    let fetched = get_progress(&pool, user, &uuid, ProgressFormat::Epub)
+        .await
+        .unwrap()
+        .unwrap();
+    assert_eq!(fetched.epub_cfi, saved.epub_cfi);
+}
+
+#[tokio::test]
+async fn upsert_is_last_write_wins() {
+    let pool = init_db("sqlite::memory:").await.unwrap();
+    let user = seed_user(&pool, "alice").await;
+    let (_, uuid) = seed(&pool, "/lib", "Book A").await;
+    let first = ProgressUpdate {
+        book_uuid: uuid.clone(),
+        format: ProgressFormat::Epub,
+        epub_cfi: Some("epubcfi(/6/4!/4/2/1:0)".into()),
+        audio_position_seconds: None,
+    };
+    upsert_progress(&pool, user, &first).await.unwrap();
+    let second = ProgressUpdate {
+        book_uuid: uuid.clone(),
+        format: ProgressFormat::Epub,
+        epub_cfi: Some("epubcfi(/6/12!/4/8/3:7)".into()),
+        audio_position_seconds: None,
+    };
+    let saved = upsert_progress(&pool, user, &second).await.unwrap();
+    assert_eq!(saved.epub_cfi.as_deref(), Some("epubcfi(/6/12!/4/8/3:7)"));
+}
+
+#[tokio::test]
+async fn isolates_per_user_book_format() {
+    let pool = init_db("sqlite::memory:").await.unwrap();
+    let alice = seed_user(&pool, "alice").await;
+    let bob = seed_user(&pool, "bob").await;
+    let (_, uuid) = seed(&pool, "/lib", "Book A").await;
+    upsert_progress(
+        &pool,
+        alice,
+        &ProgressUpdate {
+            book_uuid: uuid.clone(),
+            format: ProgressFormat::Epub,
+            epub_cfi: Some("epubcfi(alice)".into()),
+            audio_position_seconds: None,
+        },
+    )
+    .await
+    .unwrap();
+    upsert_progress(
+        &pool,
+        alice,
+        &ProgressUpdate {
+            book_uuid: uuid.clone(),
+            format: ProgressFormat::Audio,
+            epub_cfi: None,
+            audio_position_seconds: Some(42.5),
+        },
+    )
+    .await
+    .unwrap();
+    // Bob has no row yet.
+    assert!(get_progress(&pool, bob, &uuid, ProgressFormat::Epub)
+        .await
+        .unwrap()
+        .is_none());
+    // Alice's two rows don't trample each other.
+    let epub = get_progress(&pool, alice, &uuid, ProgressFormat::Epub)
+        .await
+        .unwrap()
+        .unwrap();
+    let audio = get_progress(&pool, alice, &uuid, ProgressFormat::Audio)
+        .await
+        .unwrap()
+        .unwrap();
+    assert_eq!(epub.epub_cfi.as_deref(), Some("epubcfi(alice)"));
+    assert_eq!(audio.audio_position_seconds, Some(42.5));
+}
+
+#[tokio::test]
+async fn upsert_unknown_book_is_not_found() {
+    let pool = init_db("sqlite::memory:").await.unwrap();
+    let user = seed_user(&pool, "alice").await;
+    let res = upsert_progress(
+        &pool,
+        user,
+        &ProgressUpdate {
+            book_uuid: "no-such-uuid".into(),
+            format: ProgressFormat::Epub,
+            epub_cfi: Some("epubcfi(x)".into()),
+            audio_position_seconds: None,
+        },
+    )
+    .await;
+    assert!(matches!(res, Err(ProgressError::BookNotFound)));
+}
+
+#[tokio::test]
+async fn get_progress_returns_none_when_unset() {
+    let pool = init_db("sqlite::memory:").await.unwrap();
+    let user = seed_user(&pool, "alice").await;
+    let (_, uuid) = seed(&pool, "/lib", "Book A").await;
+    assert!(get_progress(&pool, user, &uuid, ProgressFormat::Epub)
+        .await
+        .unwrap()
+        .is_none());
+}
+
+#[tokio::test]
+async fn record_session_inserts_per_format_row() {
+    let pool = init_db("sqlite::memory:").await.unwrap();
+    let user = seed_user(&pool, "alice").await;
+    let (book_id, uuid) = seed(&pool, "/lib", "Book A").await;
+    record_session(
+        &pool,
+        user,
+        &SessionReport {
+            book_uuid: uuid.clone(),
+            format: ProgressFormat::Epub,
+            started_at: 100,
+            ended_at: 460,
+            progress_units: 360,
+            device_id: None,
+        },
+    )
+    .await
+    .unwrap();
+    let count: i64 = sqlx::query_scalar(
+        "SELECT COUNT(*) FROM reading_sessions WHERE user_id = ? AND book_id = ?",
+    )
+    .bind(user)
+    .bind(book_id)
+    .fetch_one(&pool)
+    .await
+    .unwrap();
+    assert_eq!(count, 1);
+
+    record_session(
+        &pool,
+        user,
+        &SessionReport {
+            book_uuid: uuid.clone(),
+            format: ProgressFormat::Audio,
+            started_at: 200,
+            ended_at: 800,
+            progress_units: 600,
+            device_id: None,
+        },
+    )
+    .await
+    .unwrap();
+    let audio_count: i64 = sqlx::query_scalar(
+        "SELECT COUNT(*) FROM listening_sessions WHERE user_id = ? AND book_id = ?",
+    )
+    .bind(user)
+    .bind(book_id)
+    .fetch_one(&pool)
+    .await
+    .unwrap();
+    assert_eq!(audio_count, 1);
+
+    // Unknown uuid → skipped, returns false so the REST handler can
+    // count only the rows that actually landed (issue: copilot review
+    // on #300).
+    let skipped = record_session(
+        &pool,
+        user,
+        &SessionReport {
+            book_uuid: "no-such-uuid".into(),
+            format: ProgressFormat::Epub,
+            started_at: 0,
+            ended_at: 10,
+            progress_units: 10,
+            device_id: None,
+        },
+    )
+    .await
+    .unwrap();
+    assert!(!skipped, "unknown uuid should be skipped (returns false)");
+}
